+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Introduced `RangeGappable` trait for collections with `range` method
+- Exported `bounds` function for extracting the bounds of a range
 
 
 0.1.0